Refactor sound check service, streams
